Removed LoadedValue and implemented "serialization".

+++ Sources/Fonts/FontCharacter.cpp
@@ -2,7 +2,7 @@
 
 namespace acid
 {
-	FontCharacter::FontCharacter(const int &id, const double &textureCoordX, const double &textureCoordY, const double &textureSizeX, const double &textureSizeY, const double &offsetX, const double &offsetY, const double &sizeX, const double &sizeY, const double &advanceX) :
+	FontCharacter::FontCharacter(const int &id, const float &textureCoordX, const float &textureCoordY, const float &textureSizeX, const float &textureSizeY, const float &offsetX, const float &offsetY, const float &sizeX, const float &sizeY, const float &advanceX) :
 		m_id(id),
 		m_textureCoordX(textureCoordX),
 		m_textureCoordY(textureCoordY),
